[WEB] Added endpoint protection
