+++ server/package.json
@@ -15,6 +15,7 @@
     "chai-http": "^4.3.0",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
+    "express-jwt": "^5.3.1",
     "jsonwebtoken": "^8.5.1",
     "mocha": "^6.2.2",
     "mongoose": "^5.7.7",
