+++ server/src/logs/combined.log
@@ -260,3 +260,110 @@ undefined [[object Object]] undefined
 undefined [[object Object]] undefined
 undefined [[object Object]] undefined
 undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
+undefined [[object Object]] undefined
