+++ server/src/routes/api/api.routes.js
@@ -1,5 +1,7 @@
 import { Router } from 'express';
-import  { getUser, addUser, authenticate, generateToken, respondToken}  from './../../controllers/apiController.controller' 
+import  { getUser, addUser }  from './../../controllers/apiController.controller' 
+import { authenticate, generateToken, respondToken } from './../../controllers/authController.controller';
+import { ensureToken, verifyToken } from './../../config/jwt';
 import bodyParser from 'body-parser';
 import { add } from 'winston';
 
@@ -12,7 +14,7 @@ router.get('/', (req, res) => {
 });
 
 router.get('/find/:uid', getUser);
-router.post('/add/', urlEncodedParser, addUser);
+router.post('/add/', urlEncodedParser, ensureToken, verifyToken, addUser);
 router.post('/auth/', urlEncodedParser, authenticate, generateToken, respondToken);
 
 export default router;
